+++ src/gateways/ServerGateway.py
@@ -5,8 +5,8 @@ from fastapi import FastAPI, WebSocket, WebSocketDisconnect
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import StreamingResponse
 
-from model.custom_types.index import MotionCommand
-from model.Robotdog import Robotdog
+from src.custom_types.index import MotionCommand
+from src.model.Robotdog import Robotdog
 
 class ServerGateway:
     def __init__(self, robotdog: Robotdog):
